Breaking change: Move `TopSecret::{BatchResult, Result}` to `TopSecret::Text::{BatchResult, Result}`
In #57 we introduced `TopSecret::FilteredText::Result` since
`TopSecret::Result` was already taken.

This commit simly keeps parity with the change, and better organizes
each result object.
