--- lib/top_secret/result.rb
@@ -1,24 +1,0 @@
-# frozen_string_literal: true
-
-module TopSecret
-  # Holds the result of a redaction operation.
-  class Result
-    # @return [String] The original unredacted input
-    attr_reader :input
-
-    # @return [String] The redacted output
-    attr_reader :output
-
-    # @return [Hash] Mapping of redacted labels to matched values
-    attr_reader :mapping
-
-    # @param input [String] The original text
-    # @param output [String] The redacted text
-    # @param mapping [Hash] Map of labels to matched values
-    def initialize(input, output, mapping)
-      @input = input
-      @output = output
-      @mapping = mapping
-    end
-  end
-end
